Refactor to move prov stuff to coreprov.c (4/4)
This is the final commit of this series. It moves the functionality
to initialize the proc_alloc_me and prov_not_alloc_me lists out of
schedule.c and into coreprov.c.

At first it was unclear if we wanted to do the initialization of these
lists here, or wait until we do initialization in our corealloc.c file
(because these lists are used in the coreallocation process as well).
However, these list only exist because provisioning exits, so it
probably makes sense to initialize them here. Anyway, the lists are only
accessed by the core allocation logic upon branching to determine that a
core is provisioned in the first place.

Signed-off-by: Kevin Klues <klueska@cs.berkeley.edu>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>