Kconfig option to change core allocation policy
We still have only option for core allocation at this point, but the
scaffolding is now in place to add others. As part of this, we rename
corealloc.c to fcfs_corealloc.c because we want to be able to add a
prefix to files that implement other core allocation strategies.

Signed-off-by: Kevin Klues <klueska@cs.berkeley.edu>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>