pthread: Account for pth stopping in has_blocked
This popped up as

	uthread.c:621: run_uthread: Assertion `uthread->state == 2' failed.

and

	pthread.c:246: pth_sched_entry: Assertion `new_thread->state == 2'
	failed.

for an epoll/eventfd app.

The ready and active queues were corrupted, due to adding a pthread to
the ready queue in pth_thread_runnable() that was still on the active
queue.

Anytime a pthread stops and will eventually have pth_thread_runnable()
called, as is the case with blocking a uthread on an event queue (which
epoll does), then the pthread code needs to yank it off the active
queue.  Modifications of pthread->state are a good sign that list
management needs to be done.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>