Use a -D when building parlib
The -D will be defined so that a parlib header can determine how to
 #include a file.

The search order for user libraries is to first check its include
directory in the source code, and then check the system headers.  This
is important so that we build with the latest headers, i.e. the ones we
just changed.

The problem comes when we want to include a system/glibc header that
happens to have the same name as a parlib header, e.g. assert.h in an
upcoming patch.  From within parlib, we could do an #include_next.  But
for external libraries and apps, that fails since user/parlib/include is
not on their search path.

With this -D, we can tell which situation we're in and #include
accordingly.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>