Fix binutils/gold breakage (XCC)
Binutils's safe-ctype.h plays some games with cpp macros that
are of dubious legalities as far as C++ is concerned: it is
not clear that the standard permits one to create macros with
the names of the various functions from <ctype.h>. In
particular, these are defined as functions in C++, as opposed
to macros.

"safe-ctype.h" #undef's these and #define's them to symbols that
are guaranteed to fail; this caused some problems under another
compilation environment that included C++ headers that defined
classes with member functions with the same names as the symbols
from <ctype.h>; these were being expanded to these failing
symbols.

The workaround is to move inclusion of safe-ctype.h to the end
of the list of includes. We should probably upstream this.

Signed-off-by: Dan Cross <crossd@gmail.com>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>