Add a set of ak-scripts for use by docopt-scripts
The docopt-scripts tool provides a framework for invoking a collection
of bash scripts through a single command line interface. Its purpose is
to ease the development of these scripts with well-formed usage
semantics from docopt and collect them into a single place for easy
access.

The docopt-scripts tool can be downloaded from here:
https://github.com/google/docopt-scripts

Signed-off-by: Kevin Klues <klueska@cs.berkeley.edu>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>