Rename static_assert to avoid conflicting with C++'11 (and later) (XCC)
The C++ standard laid claim to the name, 'static_assert' in the
C++'11 standard; our macro caused conflict trying to build C++
programs. Rename ours to avoid the conflict.

C11 also has support for _Static_assert and the 'static_assert'
macro in the 'assert.h' header. We should migrate to that.  This
is an incremental step in that direction.

Signed-off-by: Dan Cross <crossd@gmail.com>
[removed undef of parlib_static_assert; we aren't overriding stuff]
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>