Weasel apart parlib/libc symbols for signals (XCC)
Previously, all of the akaros signal handling code was in parlib, which meant
that none of the glibc functions could use it. Moreover, there was no
good way for an individul 2LS to overwrite many of the commong signal
handling functions with its own functionality.  This commit weasels
apart some of this mess, allowing a 2LS to hook in its own set of
functions for common functions for masking/relaying posix signals.
Moreover, it allows glibc to now access the generic signal_ops struct,
so any function pointers a 2LS wishes to set in there can be called from
glibc if desired.  This will prove useful when implementing SIGPIPE in a
subsequent commit.

Signed-off-by: Kevin Klues <klueska@cs.berkeley.edu>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>