Remove need for externally alloced sigdata struct
We now use space on the stack for storing the original user context and
and any signal specific state when triggering a posix signal on a
pthread.

As a side effect, we should now be able to safely longjmp out of a
signal handler because there is no longer any external state
to clean up when the signal handler completes.

Signed-off-by: Kevin Klues <klueska@cs.berkeley.edu>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>