Add a sigself() signal_op
All of the existing signal_ops map directly to traditional signal
operations fro glibc. This commit introduces a new operation, called sigself().
This operation is meant to trigger a signal immediately on the calling
thread.  Before this call returns, the signal passed into it will be
generated and run to compeltion.

Signed-off-by: Kevin Klues <klueska@cs.berkeley.edu>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>