Update the user Makefrag to be more generic
This commit updates the makefrag for user libs so that it can handle
libraries that organize their header files using different patterns. As
part of this, we are now able to remove our custom pthread Makefile, as
one of the patterns we now support is the one used by the pthread
library.

Signed-off-by: Kevin Klues <klueska@cs.berkeley.edu>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>