Use a helper to copy out the current context
The helper allows us to run some code before we copy out the current
context.  I'll use this shortly as the location to 'finalize' the
context (fully copy it out of hardware).  In the future, we might want
other things in here.

I considered making this take the cur_ctx * to avoid another core_id()
call, but most callers were already passing current_ctx.  Also, if we
didn't use current_ctx (and took a pointer) we could inline the call in
trap.h.  But it's not really worth the hassle yet.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>