Add support for sleep, nanosleep, and usleep (XCC)
The nanosleep call is a pass through to SYS_nanosleep, the usleep call
is a pass through to SYS_block, and sleep is implemented in terms of
nanosleep. We previously had our own copy of sysdeps/posix/sleep.c in
sysdeps/akaros, but this is now removed in favor of
sysdeps/posix/sleep.c which uses nanosleep under the hood.

Signed-off-by: Kevin Klues <klueska@cs.berkeley.edu>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>