Determine bufs with is_ktask in the VFS [1/2]
We had been using the existence of current to signal whether or not we want
to operate on the kernel's behalf.  Use is_ktask() instead.

This doesn't solve the overall issue with KFOP (kernel file operation).  We
should either just use waserror() protected memcpy() calls or take out the
is_user_r{w,}addr() calls from copy_to_user.  The checks on the validity of
a user's address must be done on entry (e.g. syscall.c).  It's not clear
that we need to do the checks later on, especially if it requires the
is_ktask() hassle (especially since these only pop up when there's a bug).

This will build, but all of the kernel ops will fail until the next patch.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>