x86: pair disable_irqs with enable_irqs
We were being a little sloppy and should be more explicit about our use of
enabling/disabling IRQs.  If we enable irqs at a certain point, we should
turn them back off later instead of assuming someone downstream will do it.
It's more of a whitelist approach to when IRQs are on.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>