x86: clean up __handle_page_fault()
Other than being a mess, there were a couple issues.  One change is that
IRQs are disabled for the entire trap handling.  We're in a kernel trap
handler - turning on IRQs could lead to a mess.  The other change is that
we call handle_page_fault_nofile().  If a user asks the *kernel* to operate
on an mmaped file addr, the kernel will not do this, and we'll end up
killing the process.  If there's a compelling case for handling this, then
we can add it back in later.  For now, I'm opting for safety.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>