Allow syscall-less abortable CLEs
That assert was for safety, but I think we are fine without it.  Later on,
we branch on cle->sysc anyways.

Here's a scenario that could trigger this.  I tend to put:

	kthread_usleep(1000);

in kfs_readpage() to simulate having a filesystem that could actually
block.  We take a page fault and attempt to load the page.  kfs_readpage
eventually gets called and blocks.

Recently I changed kthread_usleep() to use rendezes, which are abortable.
Now we hit this assert.

The thing is that we don't have a syscall since the user never issued one.
The page fault is an implicit syscall.

Right now, the only time this happens is when an SCP faults on a
file-backed VMR.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>