Refactors backtrace functions [1/2]
We had a mess of functionality with repeated code in various spots, which
made it difficult to add new features.

As part of this, we have backtrace_hwtf(), which can print out the
appropriate BT based on whether the TF is the kernel's or the user's.  This
means that CTRL-B will work on user TFs, and it will spit out text (via
printk) that can be processed with scripts/bt-akaros.sh.

So now the answer to "how do I backtrace a deadlocked user SCP" is
"CTRL-B", instead of a long discussion.  =)

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>