Do not call printf from sigaction (XCC)
Right now, it's okay, since parlib redefines printf to akaros_printf.  That
tricks whatever dark magic that triggers the 'multiple libcs' problem.  But
when we fix printf to not use a macro-based solution, this breaks.

Until we understand more, we need to continue to be careful about where we
call printf or similar functions within glibc.  Normally, we'll find out
when the toolchain build fails, but with the #undef printf in
parlib/stdio.h, we were hiding those failures.  That will be fixed shortly.

Rebuild glibc.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>