x86: Detect and handle missing perf support
If a machine has perf version 0, which is the case for my Qemu, we'll get a
GPF during initialization.  The per core initialization and any accesses to
the Qperf file will abort if we don't have the right version.

This assumes that if open of a Qperf fails, that there is no other way for
the user to trigger access to the perf MSRs.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>