Clean up the PTE() macro (XCC)
The PTE macro was a bit of a mess.  Every user of it had the PA handy.  But
they always converted to the PPN.  Plus, x86's kpte_write() wasn't even
using it.

It also didn't need to be in a kernel header, which was causing some
collisions for some applications.  The mmu header is so that userspace can
be aware of their address space and walk their own page tables.  PTE() was
something that could build a page table entry, which they don't need.

I didn't compile the RISC-V stuff.  It's slowly rotting.

Reinstall your kernel headers.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>