Clean up 'devgen()' a bit.
Try to make this a tad cleaner and clearer, and be a bit more
defensive in the implementation. In particular, we try to check
the validity of table the table index and associated entry before
dereferencing, thus ensuring the function's preconditions hold.

Clean up the comments a bit, to make it less ambiguous whether
they refer to '.' or '..'. Also fix a (presumed?) grammar error.

Signed-off-by: Dan Cross <crossd@gmail.com>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>