VMM: init the vmexit counters when initing the VMM
Previously, this was just done during the first vmctl command
(REG_RSP_RIP_CR3).  It's really a "once per VMM per process" thing.  In the
future, we could add commands to reset the counters or something.  Still,
the initial initialization should be when setting up the VMM in the first
place.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>