VMM: Add a helper to emulate MSRs [1/4]
The existing msrio() is too ingrained in the KVM/Dune style of operations.
It'll work a little easier with pointers to the registers and not having a
ret value that is a VMX header value.

The existing implementation of MSR emulation is in vmx.c, though for the
most part it is VMX-independent.  The one bit that is from VMX is using the
opcodes from VMX exits to specify whether we're reading or writing.  That
will change shortly.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>