x86: Enumerate reserved eflags (XCC)
Eflags/rflags has a few reserved fields.  If you attempt a VM enter with a
reserved bit set incorrectly, vmlaunch/resume will fail due to the bad
state.  Interestingly enough, this won't happen if you attempt to popfl a
bad rflags.  So VMX is picky, but the regular hardware isn't.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>