VMM: Clean up per-cpu VMCS state
When we want to run a VM on a new host pcore, we need to set up a bunch of
per-core info, such as the stacktop, TLS, and GDT.

This is a bit of a cleanup of the old code, which needed to handle things
we didn't, such as the 32 bit SYSENTER_ESP.  It was also a classic "bag on
the side," where instead of just using the pcpui, it was reading the
hardware state (e.g. 'str') and using those values.

In the newer version, we just save whatever we know we need, using pcpui
for the source of all of the info.  In the process, we remove a lot of
unnecessary code.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>