VMM: Add custom exit reasons (XCC)
We'll need these types for when the kernel fails to pop a VM context.  In
that case, we'll reflect the context back to the 2LS, and we'll need exit
reasons to explain what went wrong.

I also cleaned up the #defines a little.  We had multiple copies of the
exit reasons throughout the codebase.

Reinstall your kernel headers.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>