VMM: Add kernel support for VM contexts [1/2]
The kernel now knows how to pop VM contexts and handle VM exits.

As of this commit, we're still using the old KVM loop.  The HOST_RIP on
resume is still set to use the old KVM loop, IPI injection still uses the
vmctl, and userspace does not ask it to use contexts.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>